feat: update prototype
